requirements.txt
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
defusedxml==0.6.0
django-crispy-forms==1.7.2
idna==2.8
oauthlib==3.0.1
PyJWT==1.7.1
python-dotenv==0.10.2
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.3
